backend/requirements.txt
# ============================================
# Profectus AI - Complete Dependencies
# ============================================

# ----------------- Web Scraping -----------------
httpx>=0.27
selectolax>=0.3.21
beautifulsoup4>=4.12
aiohttp>=3.9
requests>=2.31

# ----------------- Data Processing -----------------
pandas>=2.0
numpy>=1.24
jsonschema>=4.20
openpyxl>=3.1
xlsxwriter>=3.2

# ----------------- ML / Embeddings -----------------
sentence-transformers>=2.2
transformers>=4.35
torch>=2.0
scikit-learn>=1.3

# ----------------- Vector Search -----------------
faiss-cpu>=1.7
# Use faiss-gpu>=1.7 if you have CUDA support

# ----------------- LLM / Google Cloud -----------------
google-cloud-aiplatform>=1.38
google-api-core>=2.15
google-cloud-firestore>=2.14

# ----------------- API -----------------
fastapi>=0.110
uvicorn>=0.27

# ----------------- NLP -----------------
nltk>=3.8
ftfy>=6.1

# ----------------- Database -----------------
sqlite-utils>=3.36

# ----------------- YouTube -----------------
yt-dlp>=2023.0
youtube-transcript-api>=0.6

# ----------------- Google ADK -----------------
google-adk>=0.1

# ----------------- Utilities -----------------
tqdm>=4.66
pydantic>=2.8
python-dotenv>=1.0
pyyaml>=6.0.2
rich>=13.7
